utilites for creating interpreter
